:hammer: Fix: createSeriesReview API & series.findOne

+++ backend-server/src/apis/seriesReviews/seriesReviews.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { SeriesReview } from './entities/seriesReviews.entity';
 import { Repository } from 'typeorm';
@@ -6,35 +6,49 @@ import {
   ISeriesReviewsServiceCreate,
   ISeriesReviewsServiceDelete,
   ISeriesReviewsServiceUpdate,
-  ISeriesReviewsServicefindBySeries,
-  ISeriesReviewsServicefindOne,
+  ISeriesReviewsServiceFindBySeries,
+  ISeriesReviewsServiceFindOne,
+  ISeriesReviewsServiceFindSeriesRating,
+  ISeriesReviewsServiceFindBySeriesAndUser,
 } from './interfaces/seriesReviews-service.interface';
+import { PaymentDetailsService } from '../paymentDetails/paymentDetails.service';
 
 @Injectable()
 export class SeriesReviewsService {
   constructor(
     @InjectRepository(SeriesReview)
     private readonly seriesReviewRepository: Repository<SeriesReview>,
+
+    private readonly paymentDetailsService: PaymentDetailsService,
   ) {}
 
   findAll(): Promise<SeriesReview[]> {
     return this.seriesReviewRepository.find({ relations: ['series', 'user'] });
   }
 
-  findOne({ reviewId }: ISeriesReviewsServicefindOne): Promise<SeriesReview> {
+  findOne({ reviewId }: ISeriesReviewsServiceFindOne): Promise<SeriesReview> {
     return this.seriesReviewRepository.findOne({ where: { reviewId } });
   }
 
   findBySeries({
     seriesId,
-  }: ISeriesReviewsServicefindBySeries): Promise<SeriesReview[]> {
+  }: ISeriesReviewsServiceFindBySeries): Promise<SeriesReview[]> {
     return this.seriesReviewRepository.find({
       where: { series: { seriesId } },
       relations: ['series', 'user'],
     });
   }
+  findBySeriesAndUser({
+    seriesId,
+    user,
+  }: ISeriesReviewsServiceFindBySeriesAndUser): Promise<SeriesReview> {
+    return this.seriesReviewRepository.findOne({
+      where: { series: { seriesId }, user: { userId: user.userId } },
+      relations: ['series', 'user'],
+    });
+  }
 
-  async findSeriesRating({ seriesId }: ISeriesReviewsServicefindBySeries) {
+  async findSeriesRating({ seriesId }: ISeriesReviewsServiceFindSeriesRating) {
     const result = await this.seriesReviewRepository
       .createQueryBuilder('review')
       .select('Avg(review.rating)', 'rating')
@@ -47,30 +61,62 @@ export class SeriesReviewsService {
     return result.rating;
   }
 
-  create({
+  async create({
     createSeriesReviewInput,
     user,
   }: ISeriesReviewsServiceCreate): Promise<SeriesReview> {
     const { seriesId, ...rest } = createSeriesReviewInput;
+
+    const payment = await this.paymentDetailsService.findOne({
+      seriesId,
+      user,
+    });
+    if (!payment) {
+      throw new Error('No permissions');
+    }
+
+    const review = await this.findBySeriesAndUser({
+      seriesId,
+      user,
+    });
+
+    if (review) {
+      throw new Error('Review already exists');
+    }
+
     return this.seriesReviewRepository.save({
       ...rest,
       series: { seriesId },
-      user: { userId: user },
+      user: { userId: user.userId },
     });
   }
 
   async update({
+    user,
     reviewId,
     updateSeriesReviewInput,
   }: ISeriesReviewsServiceUpdate): Promise<SeriesReview> {
     const review = await this.findOne({ reviewId });
+
+    if (review && review.user.userId !== user.userId) {
+      throw new UnauthorizedException();
+    }
+
     return this.seriesReviewRepository.save({
       ...review,
       ...updateSeriesReviewInput,
     });
   }
 
-  async delete({ reviewId }: ISeriesReviewsServiceDelete): Promise<boolean> {
+  async delete({
+    reviewId,
+    user,
+  }: ISeriesReviewsServiceDelete): Promise<boolean> {
+    const review = await this.findOne({ reviewId });
+
+    if (review && review.user.userId !== user.userId) {
+      throw new UnauthorizedException();
+    }
     const result = await this.seriesReviewRepository.delete({ reviewId });
     return result.affected ? true : false;
   }
